sclang refactoring: split into more files and reformat some comments
